+++ src/app/components/AddQuestions/AddQuestions.module.css
@@ -1,0 +1,22 @@
+.outerWrapper {
+    height: 50%;
+    margin: 0em;
+  }
+  
+  .scrollableContent {
+    height: 100%;
+    overflow-y: auto;
+    padding: 0em;
+    margin: 0em;
+  }
+  .container {
+    display: inline-block;
+    margin: 20px;
+    padding: 20px;
+    border: 1px solid #ccc;
+    border-radius: 8px;
+  }
+
+  .question {
+    padding-bottom: 20px
+  }
