+++ src/app/page.tsx
@@ -6,11 +6,9 @@ import Login from './Login';
 import AddQuestions from "@components/AddQuestions/AddQuestions"
 import { IFormQuestion } from '@database/volunteerFormSchema'
 
-
 export default function Home() {
   const [apiResponse, setApiResponse] = useState('');
   const { data, status } = useSession();
-  const [questions, setQuestions] = useState<IFormQuestion[]>([])
 
   const handleApiCall = async () => {
     try {
@@ -23,16 +21,6 @@ export default function Home() {
     }
   };
 
-  const addQuestion = () => {
-    // Create an empty question and append it to the questions array
-    const emptyQuestion: IFormQuestion = {
-      question: '',
-      fieldType: 'Multiple Choice', // Set default fieldType or adjust as needed
-      options: [],
-    };
-    setQuestions((prevQuestions) => [...prevQuestions, emptyQuestion]);
-  };
-
   return (
     <main>
       <Navbar />
@@ -43,10 +31,7 @@ export default function Home() {
         <h3>Login Website</h3>
         <Login />
       </div>
-      <AddQuestions questions={questions} setQuestions={setQuestions}/>
-      <button onClick={addQuestion}>add questions</button>
-      {//whenever button is clicked, new empty question should be appended to the list
-      }
+
     </main>
   );
 }
